requirements.txt
streamlit>=1.30,<2
python-dotenv>=1.0,<2
requests>=2.31,<3


